refactor(ballot-interpreter): make mocking the addon easier
Making `./addon` a TypeScript module means we can easily mock it.
