+++ libs/ballot-interpreter/Cargo.toml
@@ -2,7 +2,7 @@
 name = "ballot-interpreter"
 version = "0.1.0"
 edition = "2021"
-exclude = ["build/rust_addon.node"]
+exclude = ["build/addon.node"]
 
 [lib]
 crate-type = ["cdylib", "rlib"]
