+++ libs/ballot-interpreter/package.json
@@ -13,7 +13,7 @@
     "benchmark": "jest --runInBand --projects benchmarks",
     "build": "pnpm --filter $npm_package_name... build:self",
     "build:self": "pnpm install:rust-addon && pnpm build:rust-addon && pnpm build:ts",
-    "build:rust-addon": "cargo-cp-artifact -nc build/hmpb-ts/rust_addon.node -- cargo build --message-format=json-render-diagnostics --release --offline",
+    "build:rust-addon": "cargo-cp-artifact -nc build/addon.node -- cargo build --message-format=json-render-diagnostics --release --offline",
     "build:ts": "tsc --build tsconfig.build.json",
     "clean": "pnpm --filter $npm_package_name... clean:self",
     "clean:self": "cargo clean --release --package ballot-interpreter && rm -rf build && tsc --build --clean tsconfig.build.json",
