+++ libs/ballot-interpreter/src/hmpb-ts/addon.ts
@@ -1,6 +1,21 @@
 import { Election } from '@votingworks/types';
 import { ImageData } from 'canvas';
+import { createRequire } from 'node:module';
+import { join } from 'node:path';
 import { type TemplateGridAndBubbles } from './find_template_grid_and_bubbles';
+
+const addon = (() => {
+  // NOTE: this only works because the build output can get to the root of the
+  // project in the same number of `../` as the source input:
+  //
+  //   src/hmpb-ts/addon.ts -> build/addon.node via `../../build/addon.node`
+  //   build/hmpb-ts/addon.js -> build/addon.node via `../../build/addon.node`
+  //
+  const require = createRequire(__filename);
+  const root = join(__dirname, '../..');
+  // eslint-disable-next-line import/no-dynamic-require
+  return require(join(root, 'build', 'addon.node'));
+})();
 
 /**
  * The result of calling `interpret`.
@@ -32,7 +47,16 @@ export function interpret(
     scoreWriteIns?: boolean;
     disableVerticalStreakDetection?: boolean;
   }
-): BridgeInterpretResult;
+): BridgeInterpretResult {
+  return addon.interpret(
+    election,
+    ballotImageSourceSideA,
+    ballotImageSourceSideB,
+    debugBasePathSideA,
+    debugBasePathSideB,
+    options
+  );
+}
 
 /**
  * Type of the Rust `findTemplateGridAndBubbles` implementation.
@@ -40,9 +64,16 @@ export function interpret(
 export function findTemplateGridAndBubbles(
   ballotImageSourceSideA: string | ImageData,
   ballotImageSourceSideB: string | ImageData
-): TemplateGridAndBubbles;
+): TemplateGridAndBubbles {
+  return addon.findTemplateGridAndBubbles(
+    ballotImageSourceSideA,
+    ballotImageSourceSideB
+  );
+}
 
 export function runBlankPaperDiagnostic(
   image: string | ImageData,
   debugBasePath?: string
-): boolean;
+): boolean {
+  return addon.runBlankPaperDiagnostic(image, debugBasePath);
+}
