+++ libs/ballot-interpreter/src/hmpb-ts/diagnostic.ts
@@ -1,4 +1,4 @@
-import { runBlankPaperDiagnostic as runBlankPaperDiagnosticImpl } from './rust_addon';
+import { runBlankPaperDiagnostic as runBlankPaperDiagnosticImpl } from './addon';
 
 /**
  * Runs a diagnostic on a blank paper image to determine if it is a valid
