+++ libs/ballot-interpreter/src/hmpb-ts/find_template_grid_and_bubbles.ts
@@ -1,7 +1,7 @@
 import { ImageData } from 'canvas';
 import { Result, err, ok } from '@votingworks/basics';
 import { SheetOf } from '@votingworks/types';
-import { findTemplateGridAndBubbles as findTemplateGridAndBubblesImpl } from './rust_addon';
+import { findTemplateGridAndBubbles as findTemplateGridAndBubblesImpl } from './addon';
 import {
   BallotPageTimingMarkMetadata,
   Point,
