--- libs/ballot-interpreter/src/hmpb-ts/rust_addon.node
@@ -1,1 +1,0 @@
-../../build/hmpb-ts/rust_addon.node
